+++ dkutils/ssh/remote_client.py
@@ -6,7 +6,7 @@ from paramiko import SSHClient, AutoAddPolicy
 from paramiko.auth_handler import AuthenticationException
 from scp import SCPClient, SCPException
 
-LOGGER = logging.getLogger()
+from ..validation import ensure_pathlib
 
 CommandResult = collections.namedtuple('CommandResult', 'status stdin stdout stderr')
 """A namedtuple containing the results of command execution
@@ -27,16 +27,10 @@ stderr: file-like
 
 class RemoteClient:
 
-    def __init__(self, host, user, password=None, remote_path=None, key_filename=None):
-        self._host = host
-        self._user = user
-        self._password = password
-        self._remote_path = remote_path
-        self._client = None
-        self._scp = None
-        self._conn = None
-        self._key_filename = key_filename
-        """Client to interact with a remote host via SSH & SCP.
+    def __init__(self, host, user, password=None, key_filename=None, logger=None):
+        """
+        Client to interact with a remote host via SSH & SCP.
+
         Parameters
         ----------
         host: str
@@ -45,12 +39,19 @@ class RemoteClient:
             the username to authenticate as
         password: str, optional
             used for password authentication
-        remote_path: str, optional
-            the directory to upload files to on the server
         key_filename: str, optional
             the file name of the pem file to be used for authentication
-
+        logger: Python logger, optional
+            python logger
         """
+        self._host = host
+        self._user = user
+        self._password = password
+        self._client = None
+        self._scp = None
+        self._conn = None
+        self._key_filename = key_filename
+        self._logger = logger if logger else logging.getLogger(__name__)
 
     def _connect(self):
         """
@@ -59,7 +60,7 @@ class RemoteClient:
         Raises
         ------
         AuthenticationException
-        If authentication fails
+            If authentication fails
         """
         if self._client is None:
             try:
@@ -76,10 +77,10 @@ class RemoteClient:
                 self._client = client
                 self._scp = SCPClient(self._client.get_transport())
             except AuthenticationException as error:
-                LOGGER.error(f'Authentication failed:  {error}')
+                self._logger.error(f'Authentication failed:  {error}')
                 raise error
 
-    def execute_commands(self, commands):
+    def execute_commands(self, commands, stream_logs=False):
         """
         Execute multiple commands in succession.
 
@@ -87,10 +88,12 @@ class RemoteClient:
         ----------
         commands : List(str)
             List of commands as strings
+        stream_logs: Boolean, optional
+            Stream logs to python logger - this exhausts stdout
 
         Returns
         -------
-            list of CommandResult
+        list of CommandResult
 
         Examples
         --------
@@ -111,52 +114,107 @@ class RemoteClient:
         self._connect()
         results = []
         for command in commands:
-            stdin, stdout, stderr = self._client.exec_command(command)
+            if stream_logs:
+                stdin, stdout, stderr = self._client.exec_command(command, get_pty=True)
+                for line in iter(lambda: stdout.readline(2048), ""):
+                    self._logger.info(line.rstrip())
+            else:
+                stdin, stdout, stderr = self._client.exec_command(command)
             exit_code = stdout.channel.recv_exit_status()
             result = CommandResult(exit_code, stdin=stdin, stdout=stdout, stderr=stderr)
             results.append(result)
         return results
 
-    def bulk_upload(self, files):
+    def bulk_upload(self, remote_path, files):
         """
         Upload multiple files to a remote directory.
 
         Parameters
         ----------
+        remote_path: str or pathlib.PurePath
+            The directory to upload files to on the server
         files : List(str)
-            List of local files to be uploaded
-
+            List of local file paths to be uploaded
         """
+        remote_path = ensure_pathlib(remote_path)
         self._connect()
-        uploads = [self.__upload_single_file(file) for file in files]
-        LOGGER.debug(
-            f'Finished uploading {len(uploads)} files to {self._remote_path} on {self._host}'
+        for file in files:
+            self.__upload_single_file(remote_path, file)
+        self._logger.debug(
+            f'Finished uploading {len(files)} files to {remote_path} on {self._host}'
         )
 
-    def __upload_single_file(self, file):
+    def __upload_single_file(self, remote_path, file):
         """
         Upload a single file to a remote directory.
 
         Parameters
         ----------
+        remote_path: pathlib.PurePath
+            Remote path where file is uploaded
         file: str
-        The name of the local file to be uploaded
+            Local path to file being uploaded
 
         Raises
         ------
         SCPException
-        If an exception occurs uploading the file
-
+            If an exception occurs uploading the file
         """
-        upload = None
         try:
-            self._scp.put(file, recursive=True, remote_path=self._remote_path)
-            upload = file
+            self._scp.put(file, recursive=True, remote_path=str(remote_path))
         except SCPException as error:
-            LOGGER.error(error)
+            self._logger.error(error)
             raise error
-        LOGGER.debug(f'Uploaded {file} to {self._remote_path}')
-        return upload
+        self._logger.debug(f'Uploaded {file} to {remote_path}')
+
+    def bulk_download(self, remote_path, files, local_path=''):
+        """
+        Download multiple files from remote directory.
+
+        Parameters
+        ----------
+        remote_path: str or pathlib.PurePath
+            Target directory on the server from which to download files
+        files : List(str)
+            List of remote filenames to be downloaded
+        local_path : str or pathlib.PurePath, optional
+            Local destination directory of downloaded files
+        """
+        remote_path = ensure_pathlib(remote_path)
+        local_path = ensure_pathlib(local_path)
+        if not local_path.is_dir():
+            raise NotADirectoryError(f'Local path is not a directory: {local_path}')
+        self._connect()
+        for file in files:
+            self.__download_single_file(remote_path / file, local_path)
+        self._logger.debug(
+            f'Finished downloading {len(files)} files from {remote_path} on {self._host}'
+        )
+
+    def __download_single_file(self, remote_path, local_path):
+        """
+        Download a single file from a remote directory.
+
+        Parameters
+        ----------
+        remote_path: pathlib.PurePath
+            Path of the remote file or directory to be downloaded
+        local_path: pathlib.PurePath
+            Local destination directory of downloaded file
+
+        Raises
+        ------
+        SCPException
+            If an exception occurs downloading the file/directory
+        """
+        try:
+            self._scp.get(str(remote_path), local_path=str(local_path), recursive=True)
+        except SCPException as error:
+            self._logger.error(error)
+            raise error
+        self._logger.debug(
+            f'Downloaded {remote_path.name} from {remote_path.parent} to {local_path}'
+        )
 
     def disconnect(self):
         """