+++ docs/release_notes.rst
@@ -1,9 +1,22 @@
 Release Notes
 =============
 
-v0.16.0
+v1.1.0
 -------
 * Added Alteryx Gallery API client
+
+v1.0.1
+------
+* Bugfix RemoteClient bulk upload to resolve remote path issue
+
+v1.0.0
+------
+* Add bulk download method to RemoteClient
+* Move remote path from RemoteClient constructor to bulk upload & download methods
+
+v0.16.0
+-------
+* Add option to stream logs to RemoteClient commands execution
 
 v0.15.0
 -------